+++ make/mcu/STM32F7.mk
@@ -171,6 +171,7 @@ MCU_COMMON_SRC = \
             drivers/adc_stm32f7xx.c \
             drivers/audio_stm32f7xx.c \
             drivers/bus_i2c_hal.c \
+            drivers/bus_i2c_hal_init.c \
             drivers/dma_stm32f7xx.c \
             drivers/light_ws2811strip_hal.c \
             drivers/transponder_ir_io_hal.c \
